requirements.txt
paho-mqtt==1.6.1
aiocoap==0.4.1
cryptography==39.0.1
scikit-learn==1.2.2
numpy==1.24.3
scapy==2.5.0
tensorflow==2.11.0
flask==2.2.2
flask-wtf==1.0.1
